Update SenseUI-ComboChart.es6
I have replaced and redesigned the legend implementation because it did not work with the Internet Explorer because of the 'foreignObject' SVG element. I have also added the left and right margin to the legend alignment. So that the legend cannot be 'overwritten', I have also added a white background to the legend.